+++ examples/hello_world.rs
@@ -4,6 +4,7 @@ fn main() {
     let mut logger = logkit::Logger::new(Some(&logkit::StdoutTarget));
     logger.mount(logkit::LevelPlugin);
     logger.mount(logkit::TimePlugin::from_millis());
+    logger.mount(logkit::SourcePlugin);
     logkit::set_default_logger(logger);
 
     trace!("hello, this is a trace log");